requirements.txt
streamlit
pandas
xgboost
mlflow
upstash-redis
python-dotenv
scikit-learn
kaggle
prophet
tensorflow
joblib
plotly
chromadb
sentence-transformers
huggingface-hub
groq
streamlit-extras
google-generativeai
tf-keras